+++ apps/backend/sentry.server.config.ts
@@ -3,6 +3,7 @@
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
 import * as Sentry from "@sentry/nextjs";
+import { nicify } from "@stackframe/stack-shared/dist/utils/strings";
 
 Sentry.init({
   dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
@@ -14,4 +15,26 @@ Sentry.init({
   debug: false,
 
   enabled: process.env.NODE_ENV !== "development" && !process.env.CI,
+
+  // Add exception metadata to the event
+  beforeSend(event, hint) {
+    const error = hint.originalException;
+    let nicified;
+    try {
+      nicified = nicify(error);
+    } catch (e) {
+      nicified = `Error occurred during nicification: ${e}`;
+    }
+    if (error instanceof Error) {
+      event.extra = {
+        ...event.extra,
+        cause: error.cause,
+        errorProps: {
+          ...error,
+        },
+        nicifiedError: nicify(error),
+      };
+    }
+    return event;
+  },
 });