+++ apps/e2e/tests/backend/endpoints/api/v1/auth/oauth/authorize.test.ts
@@ -1,41 +1,24 @@
-import { throwErr } from "@stackframe/stack-shared/dist/utils/errors";
-import { it, localRedirectUrl } from "../../../../../../helpers";
-import { backendContext, niceBackendFetch } from "../../../../../backend-helpers";
-
-function getAuthorizeQuery() {
-  const projectKeys = backendContext.value.projectKeys;
-  if (projectKeys === "no-project") throw new Error("No project keys found in the backend context");
-
-  return {
-    client_id: projectKeys.projectId,
-    client_secret: projectKeys.publishableClientKey ?? throwErr("No publishable client key found in the backend context"),
-    redirect_uri: localRedirectUrl,
-    scope: "legacy",
-    response_type: "code",
-    state: "this is some state",
-    grant_type: "authorization_code",
-    code_challenge: "some-code-challenge",
-    code_challenge_method: "plain",
-  };
-}
+import { it } from "../../../../../../helpers";
+import { Auth, niceBackendFetch } from "../../../../../backend-helpers";
 
 it("should redirect the user to the OAuth provider with the right arguments", async ({ expect }) => {
-  const response = await niceBackendFetch("/api/v1/auth/oauth/authorize/facebook", {
-    redirect: "manual",
-    query: {
-      ...getAuthorizeQuery(),
-    },
-  });
-  expect(response.status).toBe(307);
-  expect(response.headers.get("location")).toMatch(/^http:\/\/localhost:8107\/auth\?.*$/);
-  expect(response.headers.get("set-cookie")).toMatch(/^stack-oauth-inner-[^;]+=[^;]+; Path=\/; Expires=[^;]+; Max-Age=\d+;( Secure;)? HttpOnly$/);
+  const response = await Auth.OAuth.authorize();
+  expect(response.authorizeResponse.status).toBe(307);
+  expect(response.authorizeResponse.headers.get("location")).toMatch(/^http:\/\/localhost:8107\/auth\?.*$/);
+  expect(response.authorizeResponse.headers.get("set-cookie")).toMatch(/^stack-oauth-inner-[^;]+=[^;]+; Path=\/; Expires=[^;]+; Max-Age=\d+;( Secure;)? HttpOnly$/);
+});
+
+it("should be able to fetch the inner callback URL by following the OAuth provider redirects", async ({ expect }) => {
+  const { innerCallbackUrl } = await Auth.OAuth.getInnerCallbackUrl();
+  expect(innerCallbackUrl.origin).toBe("http://localhost:8102");
+  expect(innerCallbackUrl.pathname).toBe("/api/v1/auth/oauth/callback/facebook");
 });
 
 it("should fail if an invalid client_id is provided", async ({ expect }) => {
   const response = await niceBackendFetch("/api/v1/auth/oauth/authorize/facebook", {
     redirect: "manual",
     query: {
-      ...getAuthorizeQuery(),
+      ...await Auth.OAuth.getAuthorizeQuery(),
       client_id: "some-invalid-client-id",
     },
   });
@@ -59,19 +42,44 @@ it("should fail if an invalid client_secret is provided", async ({ expect }) => 
   const response = await niceBackendFetch("/api/v1/auth/oauth/authorize/facebook", {
     redirect: "manual",
     query: {
-      ...getAuthorizeQuery(),
+      ...await Auth.OAuth.getAuthorizeQuery(),
       client_secret: "some-invalid-client-secret",
     },
   });
   expect(response).toMatchInlineSnapshot(`
     NiceResponse {
-      "status": 404,
+      "status": 401,
       "body": {
-        "code": "API_KEY_NOT_FOUND",
-        "error": "API key not found.",
+        "code": "INVALID_PUBLISHABLE_CLIENT_KEY",
+        "details": { "project_id": "internal" },
+        "error": "The publishable key is not valid for the project \\"internal\\". Does the project and/or the key exist?",
       },
       "headers": Headers {
-        "x-stack-known-error": "API_KEY_NOT_FOUND",
+        "x-stack-known-error": "INVALID_PUBLISHABLE_CLIENT_KEY",
+        <some fields may have been hidden>,
+      },
+    }
+  `);
+});
+
+it("should fail if an invalid redirect URL is provided", async ({ expect }) => {
+  const response = await niceBackendFetch("/api/v1/auth/oauth/authorize/facebook", {
+    redirect: "manual",
+    query: {
+      ...await Auth.OAuth.getAuthorizeQuery(),
+      redirect_uri: "this is an invalid URL string",
+    },
+  });
+  expect(response).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 400,
+      "body": {
+        "code": "SCHEMA_ERROR",
+        "details": { "message": "Request validation failed on GET /api/v1/auth/oauth/authorize/facebook:\\n  - Invalid URL" },
+        "error": "Request validation failed on GET /api/v1/auth/oauth/authorize/facebook:\\n  - Invalid URL",
+      },
+      "headers": Headers {
+        "x-stack-known-error": "SCHEMA_ERROR",
         <some fields may have been hidden>,
       },
     }