+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/oauth-mock-server
 
+## 2.5.16
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.5.16
+
+## 2.5.15
+
+### Patch Changes
+
+- Webhooks
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.15
+
 ## 2.5.14
 
 ### Patch Changes