+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-sc
 
+## 2.5.16
+
+## 2.5.15
+
+### Patch Changes
+
+- Webhooks
+
 ## 2.5.14
 
 ## 2.5.13